store topics of user and add scoring fix #1125
Former-commit-id: 218c9fb8352baa03ce9045107618b940e0d08edc [formerly 635f524a90fb15dc7da177c328110e0a66a0a68e] [formerly e6a37f679aa52d8eaa35b5ad31d9461a0410bc1d [formerly 982fc52096afa1cd8e7a523641ec550c0373ecb9]]
Former-commit-id: 146d6bb86827a58b093472a7ec2fd0353213f941 [formerly c9e09a58ebc0bbc0ae6f562030d15f1fc3e57917]
Former-commit-id: 0cc04e92063cb5a21fc7329922f6853e16772224

+++ iomessages.py
@@ -224,7 +224,7 @@ class Topic_Comparaison_Schema(messages.Message):
 class Scoring_Topics_Schema(messages.Message):
     topic=messages.StringField(1)
     score=messages.FloatField(2)
-    value=messages.IntegerField(3)
+    value=messages.FloatField(3)
 class Topics_Schema(messages.Message):
     items=messages.MessageField(Scoring_Topics_Schema,1,repeated=True)
     score_total=messages.FloatField(2)
